Warning
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@40370 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/unicode_util.cpp
@@ -1233,7 +1233,7 @@ ULONG UnicodeUtil::Utf16Collation::canonical(ULONG srcLen, const USHORT* src, UL
 			icu->utransClose(trans);
 
 			len *= sizeof(USHORT);
-			if (len > dstLen)
+			if (ULONG(len) > dstLen)
 				len = INTL_BAD_STR_LENGTH;
 
 			srcLen = len;
